part 5

+++ part 4/bloglist/controllers/blogs.js
@@ -5,7 +5,7 @@ const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
   
-  const blogs = await Blog.find({}).populate('user')
+  const blogs = await Blog.find({}).populate('user_Id')
   response.json(blogs)
 })
 
@@ -18,9 +18,10 @@ const getTokenFrom = request => {
 }
   
 blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
+  let blog = new Blog(request.body)
   const token = getTokenFrom(request)
   const decodedToken = jwt.verify(token, process.env.SECRET)
+  blog.user_Id=decodedToken.id
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token missing or invalid' })
   }
@@ -34,25 +35,33 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const token = getTokenFrom(request)
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
 
-blogsRouter.put('/:id', async (request, response, next) => {
+blogsRouter.put('/:id', async (request, response) => {
   const body = request.body
-
+  const token = getTokenFrom(request)
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
   const blog = {
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.number
+    likes: body.likes
   }
 
-  updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
-    
-  response.json(updatedBlog)
+  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
+  await updatedBlog.save()
+  response.status(204).json(updatedBlog)
 })
-
 blogsRouter.get('/:id', async (request, response) => {
   const blog = await Blog.findById(request.params.id)
   if (blog) {
